working nn with circular buffer of previous positions

+++ neural.h
@@ -2,4 +2,4 @@
 #include <vector>
 
 void initialize_net();
-double run_nn(std::vector<double>& in, double error);
+double run_nn(std::vector<float>& in, double error);
